Phase 6.3: extract HTTP parse helpers + TimeRange into server/httpserver
The detection/internal/operator handler shipped its own
parseTimeRange / parseIntParam / parseInt64Param. As cross-context
operator endpoints proliferate (response, future contexts), every
one would need the same helpers.

Move them into server/httpserver/params.go so they're available to
every operator handler. Lift the TimeRange{FromNs, ToNs int64} value
type alongside (it's a generic two-int64 envelope, not a detection
concept), with detection/api/types.go aliasing back via
`type TimeRange = httpserver.TimeRange` so the public api surface
stays byte-identical.

httpserver imports nothing from any bounded context, so the modular
monolith dependency graph stays one-way: contexts may import
httpserver; httpserver may not import any context. detection's
operator handler now uses httpserver.ParseTimeRange /
ParseIntParam / ParseInt64Param directly.
